+++ src/error_handling/error_messages.rs
@@ -1,3 +1,9 @@
 //! Contains constants that point to error messages
 
 pub const LEX_INCOMPLETE_STRING: u32 = 0;
+pub const LEX_INVALID_HEX_NUMBER: u32 = 1;
+pub const LEX_INVALID_OCTAL_NUMBER: u32 = 2;
+pub const LEX_INVALID_BINARY_NUMBER: u32 = 3;
+pub const LEX_INVALID_FLOATING_NUMBER: u32 = 4;
+pub const LEX_INVALID_SCIENTIFIC_NUMBER: u32 = 5;
+pub const LEX_INCOMPLETE_MULTILINE_COMMENT: u32 = 6;