+++ src/error_handling/syntax_error.rs
@@ -33,7 +33,7 @@ impl PrintableError for SyntaxError {
             )
             .finish();
 
-        report.eprint(("sample.thp", Source::from(source)));
+        report.eprint(("sample.thp", Source::from(source))).unwrap();
     }
 }